+++ Release Application/CMakeFiles/RoSA_Monitor_autogen.dir/cmake_clean.cmake
@@ -2,6 +2,7 @@ file(REMOVE_RECURSE
   "CMakeFiles/RoSA_Monitor_autogen"
   "RoSA_Monitor_autogen/include/ui_mainwindow.h"
   "RoSA_Monitor_autogen/include/ui_rosawindow.h"
+  "RoSA_Monitor_autogen/include/ui_selectrosacomponents.h"
   "RoSA_Monitor_autogen/include/ui_shelloutputwindow.h"
   "RoSA_Monitor_autogen/include/ui_simulationwindow.h"
   "RoSA_Monitor_autogen/mocs_compilation.cpp"